requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.12.0
tensorboard>=2.11.0
pyyaml>=6.0
tqdm>=4.64.0
scikit-learn>=1.2.0